Improve RequestException for network error
Currently, if a network error occurs, an unhelpful message is printed:
    RequestException: GET 0  https://taxonomy-dot-wf-xs-proto.appspot.com/CIPC-IFRS/2016/api/v1/namespace_map
    Instance of 'n_'

The "0  " occurs because HttpRequest.status and statusText are not
initialized when onError occurs.  Avoid initializing the response in
this case since it is not meaningful.

Additionally, the "Instance of 'n_'" is actually ProgressEvent, which
does not define toString().  Continue to pass this event object as the
error in case some code was actually using it, but define an additional
error formatter to give a meaningful message.

The result is:
    RequestException: GET https://taxonomy-dot-wf-xs-proto.appspot.com/CIPC-IFRS/2016/api/v1/namespace_map
    Network error.